// src/shared/types/mock-server-data.type.ts
export type MockServerData = {
  titles: string[];
  descriptions: string[];
  publicationDates: string[];
  cities: string[];
  previewImages: string[];
  images: string[];
  isPremium: boolean[];
  isFavorite: boolean[];
  types: string[];
  goods: string[];
  hostNames: string[];
  hostEmails: string[];
  hostAvatars: string[];
  hostPasswords: string[];
  hostTypes: string[];
  latitude: string[],
  longitude: string[]
};